pyproject.toml
[project]
name = "fh-heroicons"
version = "0.1.0"
description = "FastHTML package for Heroicons"
readme = "README.md"
authors = [
    { name = "Marius", email = "mariusvach@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "httpx>=0.28.0",
    "lxml>=5.3.0",
    "python-fasthtml>=0.10.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
